fix(ckdoge): use legacy P2PKH transactions for withdrawals (#8187)
Fix the withdrawal of ckDOGE to DOGE. Transactions created by the ckDOGE
minter are quite different to the ones from the ckBTC minter:

1. They use version 1 because Dogecoin does not support
[BIP-68](https://github.com/bitcoin/bips/blob/master/bip-0068.mediawiki).
2. They use the legacy P2PKH format because Dogecoin does not support
segregated witness
[BIP-141](https://github.com/bitcoin/bips/blob/master/bip-0141.mediawiki).
3. The transaction ID must be computed from the signed transaction,
since contrary to segwit transactions, the transaction ID of legacy
transactions is influenced by the signatures.

Transactions created by the ckDOGE minter are therefore widely different
from the ones created by the ckBTC minter, which requires a separate
implementation. In particular, the choice here is to use the Rust
bitcoin library to create such transactions instead of tweaking the
custom implementation used by the ckBTC minter.

Those problems also show the insufficiency of the current integration
tests, which only mock the Dogecoin canister and therefore do not ensure
the validity of the transactions created by the minter. This part is
left for future work. The current PR was tested locally on Regtest and
with the ckDOGE staging canister which successfully produced the
transaction
[32d24dcb](https://blockchair.com/dogecoin/transaction/32d24dcb68fae3cac41caa55c9f9ed39eb4ee21689ba4d989c53df243b3b7364).